litlan_bakendi/requirements.txt
asgiref==3.8.1
asyncpg==0.30.0
Django==5.1.6
djangorestframework==3.15.2
lxml==5.3.1
marisa-trie==1.2.1
numpy==2.1.3
sqlparse==0.5.3
psycopg2-binary>=2.9.6,<3.0.0
gunicorn>=20.1.0,<21.0.0
dj-database-url>=1.2.0,<2.0.0